fixing review re-rendered the whole component
