+++ frontend/src/pages/activityDetailPage/ActyDetailDisplay.tsx
@@ -1,40 +1,25 @@
-import { Box, Button, Grid, Typography } from '@mui/material';
-import { useState } from 'react';
+import { Box, Button, Grid } from '@mui/material';
 import { Carousel } from 'react-responsive-carousel';
 import 'react-responsive-carousel/lib/styles/carousel.min.css';
 import '../../assets/detailImageDisplay.css';
 import { TActyDetail } from '../../services/getActyById';
 import ActyInfoItem from './ActyInfoItem';
-import ActyReviews from './ActyReviews';
 import ImageItem from './ImageItem';
-import LeaveReview from './LeaveReview';
+import Reviews from './Reviews';
 // import MapDisplay from './MapDisplay';
 import ArrowBackIcon from '@mui/icons-material/ArrowBack';
 
 import { Link } from 'react-router-dom';
 import SingleMarkerMap from '../../components/SingleMarkerMap';
+
 interface ActyDetailDisplayProps {
   actyData: TActyDetail;
-  onRatingChanged: (isRatingChanded: boolean) => void;
 }
-// TODO: style image and detail side by side
 
 export default function ActyDetailDisplay(props: ActyDetailDisplayProps) {
   // const { auth } = useAuth() as IAuthContext;
   // console.log('file: ActyDetailDisplay.tsx:19 ~ ActyDetailDisplay ~ auth:', auth);
   const actyDetail = props.actyData;
-  const isRating = props.onRatingChanged;
-  const [reviewsToPrint, setReviewsToPrint] = useState<TActyDetail['reviews']>(actyDetail.reviews);
-
-  const onReviewAdded = (newReview: TActyDetail['reviews'][number]) => {
-    setReviewsToPrint(prevReview => [...prevReview, newReview]);
-    isRating(true);
-  };
-
-  const onReviewDeleted = (deletedReview: TActyDetail['reviews'][number]) => {
-    setReviewsToPrint(prevReview => prevReview.filter(r => r !== deletedReview));
-    isRating(true);
-  };
 
   return (
     <Grid container rowSpacing={2} columnSpacing={3}>
@@ -54,7 +39,7 @@ export default function ActyDetailDisplay(props: ActyDetailDisplayProps) {
         </Carousel>
       </Grid>
       <Grid item xs={12} md={6}>
-        <ActyInfoItem actyDetail={actyDetail} reviewTotal={reviewsToPrint.length} />
+        <ActyInfoItem actyDetail={actyDetail} reviewTotal={actyDetail.reviews.length} />
       </Grid>
 
       {/** display Mapbox */}
@@ -71,25 +56,7 @@ export default function ActyDetailDisplay(props: ActyDetailDisplayProps) {
       </Grid>
 
       <Grid item xs={12} md={6}>
-        <Typography variant="h5" margin={'20px 0 10px 0'}>
-          Reviews
-          <span style={{ fontSize: 'small', alignSelf: 'end', paddingLeft: 5, color: 'rgba(0, 0, 0, 0.6)' }}>
-            {reviewsToPrint.length > 0 ? `${reviewsToPrint.length} reviews` : `${reviewsToPrint.length} review`}
-          </span>
-        </Typography>
-        {/* // display review input text */}
-
-        <LeaveReview onReviewAdded={onReviewAdded} />
-
-        {reviewsToPrint.length > 0 ? (
-          <Box sx={{ marginTop: 4 }}>
-            <ActyReviews reviews={reviewsToPrint} onReviewDeleted={onReviewDeleted} />
-          </Box>
-        ) : (
-          <Box>
-            <h4>No reviews yet</h4>
-          </Box>
-        )}
+        <Reviews reviewList={actyDetail.reviews} />
       </Grid>
     </Grid>
     // https://mui.com/material-ui/react-grid/#interactive
